requirements_personal.txt
absl-py==0.7.0
astor==0.7.1
bleach==1.5.0
certifi==2019.3.9
cycler==0.10.0
Django==2.2.5
django-mysql==3.2.0
gast==0.2.1.post0
grpcio==1.16.1
h5py==2.9.0
html5lib==0.9999999
Jinja2==2.10.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==3.0.2
mkl-fft==1.0.10
mkl-random==1.0.2
mpld3==0.3
mysqlclient==1.3.13
numpy==1.15.4
pandas==0.24.1
patsy==0.5.1
Pillow==5.4.1
plotly==4.2.1
protobuf==3.6.1
pygpu==0.7.6
pykalman==0.9.5
pyparsing==2.3.1
pystocker==0.1.25
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
retrying==1.3.3
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
statsmodels==0.10.1
tensorboard==1.12.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tflearn==0.3.2
Theano==1.0.3+2.g3e47d39ac.dirty
tornado==5.1.1
tqdm==4.31.1
Werkzeug==0.14.1
